Test ItemArmorModif

+++ Assets/HUD and Panels/HealthBar.cs
@@ -12,7 +12,7 @@ public class HealthBar : MonoBehaviour
 
     public TextMeshProUGUI healthText;
     
-    public void SetMaxHealth(int health)
+    public void SetMaxHealth(float health)
     {
         slider.maxValue = health;
         slider.value = health;
@@ -20,7 +20,7 @@ public class HealthBar : MonoBehaviour
         fill.color = gradient.Evaluate(1f);
     }
 
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
         slider.value = health;
 
@@ -28,7 +28,7 @@ public class HealthBar : MonoBehaviour
 
     }
 
-    public void SetHealthText(int currentHealth, int maxHealth)
+    public void SetHealthText(float currentHealth, int maxHealth)
     {
         healthText.text = currentHealth.ToString() + "/" + maxHealth.ToString();
     }
