+++ Assets/Scripts/ObjectManager/Item.cs
@@ -5,7 +5,7 @@ public class Item : ScriptableObject
 {
     public string itemName = "New Item";
     public int baseValue;
-    public int percentageValue;
+    public float percentageValue;
     public Sprite icon = null; 
 
     public virtual void ChangePlayerStats(PlayerManagement player)
